Green-Guardian-Project-main/Green-Guardian-Project-main/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.1
pydantic==2.4.2
tavily-python==0.2.6
mem0-python==0.1.0
appwrite==4.0.0
openai==1.3.0
python-multipart==0.0.6
geopy==2.4.1
pandas==2.1.2
numpy==1.26.1
matplotlib==3.8.1
